// src/generated/UsageTokensDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UsageTokensDto = { input: number | null, uncachedInput: number | null, output: number | null, total: number | null, cachedInput: number | null, cacheWriteInput: number | null, };